Go back to normal Boto use with a Boto patcher
