+++ src/toil/lib/context.py
@@ -23,7 +23,6 @@ from boto.utils import get_instance_metadata
 
 from toil.lib.memoize import memoize
 from toil.lib.ec2 import UserError
-from toil.lib.botoCredentialAdapter import BotoCredentialAdapter
 
 log = logging.getLogger(__name__)
 
@@ -214,19 +213,7 @@ class Context(object):
     def __aws_connect(self, aws_module, region=None, **kwargs):
         if region is None:
             region = self.region
-            
-        if 'provider' in inspect.getargspec(aws_module.connect_to_region).keywords:
-            # If the module accepts a provider, we will give it one.
-            conn = aws_module.connect_to_region(region, provider=BotoCredentialAdapter(), **kwargs)
-        else:
-            # Otherwise, we give it a set of possibly temporary credentials.
-            # TODO: there's no way to handle expiration. Boto 2 really needs to
-            # accept a Provider everywhere and always pass it to the base
-            # AWSAuthConnection. Maybe we should consider not passing a
-            # Provider and instead monkey patching the original.
-            joinedArgs = BotoCredentialAdapter().kwargs()
-            joinedArgs.update(kwargs)
-            conn = aws_module.connect_to_region(region, **joinedArgs)
+        conn = aws_module.connect_to_region(region, **kwargs)
         if conn is None:
             raise RuntimeError("%s couldn't connect to region %s" % (
                 aws_module.__name__, region))
